consider normalized field length in scoring, remove customizable scoring
scoring now takes into account the length of the field, in relation to
the average field length. This is so, if one searches for the word
'purple', a document with text 'purple rain' scores higher than a
document with text 'little purple wild flower'.

The scoring mechanism was removed from the customization options,
because this part is still subject to change, and it's not possible at
the moment to design a proper future-proof public API for it.