make internal IDs shorter by encoding them in base 36
the internal document IDs are stored as strings, and typically occur
many times in an index, so it makes sense to keep them as short as
possible to save index space. This commit switches to base 36 (insead of
base 10) for their representation, making them shorter.

On the Divina Commedia benchmark, this leads to a 10% saving in space.