extract the ID field using extractField
This makes the ID field extraction consistent with other fields, making
it possible to customize the extraction logic.

_Note_: a test case that was asserting a special behavior for the ID
field extraction when it matches a default object property name (like
"constructor") was removed. It's better to let the user specify their
own extractField function instead of second guessing them. This is not
considered to be a breaking change, as the existing behavior was not
publicly documented, and would only take place in corner cases (when the
idField is set to a default object property name like "constructor",
which should be discouraged anyway).