Decouple serialization format from internal index data structure
Decoupling the inner implementation of the index data structure from the
serialization format makes it possible to optimize and improve the
internals without introducing backward incompatible changes to the
serialization.

Also, the new format is easier to generate in other languages, as it
does not depend on the radix tree implementation, but only on the
indexing pipeline.

Throws error in case of incompatible serialization format