Allow processTerm to expand a single term into several ones (#166)
This is useful for normalizing a single term into several.

Example:

`AC/DC` would normally be tokenized as `['ac', 'dc']`, but a user might
reasonably search for `ACDC`, which would not match. To fix this, one
could specify a `processTerm` function that, when given the term `ACDC`,
returns `['ac', 'dc']`, making the queries `AC/DC` and `ACDC`
equivalent.