Add discard method as an alternative to remove
The MiniSearch.discard method has the same visible effect of
MiniSearch.remove (both cause the document to stop appearing in search
results), but a different effect on the inverted index:

MiniSearch.remove takes as argument the full document, and removes it
from the inverted index, releasing memory immediately.

MiniSearch.discard instead only needs the document ID, and marks the
current version of the document as discarded so it is ignored in search
results. The inverted index is not modified though, so memory is not
released. Upon search, whenever a discarded ID is found, it is ignored,
and the inverted index for the search terms is cleaned up removing the
discarded documents.  This partially releases memory, but only for terms
that have been searched at least once after discarding the document.