Discard document as an alternative to remove (#183)
* Maintain mapping from ID to short ID

This speeds up removal, allows checking for duplicate IDs upon adding a
document, and prepares the ground to implement discarding documents by ID.

* Add discard method as an alternative to remove

The MiniSearch.discard method has the same visible effect of
MiniSearch.remove (both cause the document to stop appearing in search
results), but a different effect on the inverted index:

MiniSearch.remove takes as argument the full document, and removes it
from the inverted index, releasing memory immediately.

MiniSearch.discard instead only needs the document ID, and marks the
current version of the document as discarded so it is ignored in search
results. The inverted index is not modified though, so memory is not
released. Upon search, whenever a discarded ID is found, it is ignored,
and the inverted index for the search terms is cleaned up removing the
discarded documents.  This partially releases memory, but only for terms
that have been searched at least once after discarding the document.

* Implement MiniSearch.has method

To test if a document with the given ID was added to the index.

* Add cleanupDiscarded method to clean up index

This method traverses the inverted index in batches and cleans up
discarded documents.

* CLean up stored fields upon discard

* Auto vacuuming of discarded documents

* Add missing test

* Fix promise returned by vacuum when another vacuuming is in progress

* Improve documentation

* Change auto vacuum defaults and improve behavior

* Remove unnecessary default

* Fix and test a corner case with autovacuuming options

* Implement replace method

* Fix autovacuum behavior on multiple calls to discard

* Always apply broadest conditions to enqueued vacuum

* Fix issue when batch is bigger than number of documents

* Add termCount getter

* Make getDefault return autoVacuum defaults

* Fixes for auto vacuum default and batching

* Improve documentation

* Add discardAll method