Fix unwanted side effects of boost, and cascading of combineWith
The `boost` search option was mistakenly causing boosted fields to be
included in search, even when they were not in the `fields` search
option.

The `combineWith` search option was not properly taking its default from
the `SearchOptions` argument when using `search` with a `QuerySpec`
tree.

Both issues are fixed by this commit, that also adds corresponding
tests.