Add queryTerms to the search results (#241)
The `queryTerms` field is useful to know which query terms matched in
the search result. When performing an exact search, `terms` and
`queryTerms` are identical. In case of prefix or fuzzy match though,
they can be different.

For example, if one searches for "moto" using prefix search, and a
result matches with the word "motorcycle", then for that result the
`terms` array will include "motorcycle", while the `queryTerms` array
will include "moto".